refactor(shared): rename clerk-rq directory to query
