+++ packages/shared/src/react/billing/useInitializePaymentMethod.tsx
@@ -1,9 +1,9 @@
 import { useCallback, useMemo } from 'react';
 
 import type { BillingInitializedPaymentMethodResource, ForPayerType } from '../../types';
-import { defineKeepPreviousDataFn } from '../clerk-rq/keep-previous-data';
-import { useClerkQueryClient } from '../clerk-rq/use-clerk-query-client';
-import { useClerkQuery } from '../clerk-rq/useQuery';
+import { defineKeepPreviousDataFn } from '../query/keep-previous-data';
+import { useClerkQueryClient } from '../query/use-clerk-query-client';
+import { useClerkQuery } from '../query/useQuery';
 import { useOrganizationBase } from '../hooks/base/useOrganizationBase';
 import { useUserBase } from '../hooks/base/useUserBase';
 import { useBillingIsEnabled } from '../hooks/useBillingIsEnabled';
