+++ packages/shared/src/react/hooks/useOrganizationCreationDefaults.tsx
@@ -1,7 +1,7 @@
 import { eventMethodCalled } from '../../telemetry/events/method-called';
 import type { EnvironmentResource } from '../../types/environment';
-import { defineKeepPreviousDataFn } from '../clerk-rq/keep-previous-data';
-import { useClerkQuery } from '../clerk-rq/useQuery';
+import { defineKeepPreviousDataFn } from '../query/keep-previous-data';
+import { useClerkQuery } from '../query/useQuery';
 import { useAssertWrappedByClerkProvider, useClerkInstanceContext } from '../contexts';
 import { useUserBase } from './base/useUserBase';
 import { useOrganizationCreationDefaultsCacheKeys } from './useOrganizationCreationDefaults.shared';
