+++ packages/shared/src/react/query/infiniteQueryOptions.ts
