+++ packages/shared/src/react/query/queryOptions.ts
