+++ packages/shared/src/react/query/useInfiniteQuery.ts
